+++ packages/api/src/configuration/__test__/index.spec.ts
@@ -4,6 +4,8 @@ import { getConfiguration } from "../configs";
 
 const validEnv = {
   API_PORT: "123456",
+  API_SENTRY_DSN: "API_SENTRY_DSN",
+  API_SENTRY_ENVIRONMENT: "API_SENTRY_ENVIRONMENT",
   KINTO_BUCKET: "KINTO_BUCKET",
   KINTO_LOGIN: "KINTO_LOGIN",
   KINTO_PASSWORD: "KINTO_PASSWORD",
@@ -20,6 +22,8 @@ it("should return the env configuration", () => {
   expect(getConfiguration(validEnv)).toMatchInlineSnapshot(`
     Object {
       "apiPort": 123456,
+      "apiSentryDsn": "API_SENTRY_DSN",
+      "apiSentryEnvironment": "API_SENTRY_ENVIRONMENT",
       "kintoBucket": "KINTO_BUCKET",
       "kintoLogin": "KINTO_LOGIN",
       "kintoPassword": "KINTO_PASSWORD",