+++ packages/api/src/server.ts
@@ -1,9 +1,20 @@
 import * as cors from "@koa/cors";
+import * as Sentry from "@sentry/node";
 import * as Koa from "koa";
 import * as bodyParser from "koa-bodyparser";
 import { configuration } from "./configuration";
 import { router } from "./routes";
 import { logger } from "./util";
+
+if (!!configuration.apiSentryDsn) {
+  logger.info(`Logging to sentry DSN ${configuration.apiSentryDsn}`);
+  logger.info(`Sentry environment is ${configuration.apiSentryEnvironment}`);
+  Sentry.init({
+    debug: true,
+    dsn: configuration.apiSentryDsn,
+    environment: configuration.apiSentryEnvironment
+  });
+}
 
 const app = new Koa();
 
@@ -13,8 +24,6 @@ app.use(cors());
 app.use(router.routes());
 app.use(router.allowedMethods());
 
-logger.error("testing the errors", new Error("just a test"));
-logger.fatal("fatal error just a test");
 app.on("error", (err: Error, ctx: Koa.Context) => {
   logger.error(`url: ${ctx.originalUrl}: `, err);
 });