+++ packages/api/src/util/logger/logger.test.ts
@@ -1,6 +1,7 @@
 const now = 1570137016663;
 
 describe("test the output of the logger", () => {
+  const sentry = require("@sentry/node");
   let mockStdout: jest.SpyInstance;
   let logger: any;
 
@@ -12,6 +13,12 @@ describe("test the output of the logger", () => {
     mockStdout.mockImplementation(() => {});
     // Mock `Date.now()` so we have a comparable timestamp.
     global.Date.now = jest.fn(() => now);
+    // Mock `captureException` from Sentry to check that it's correctly called.
+    sentry.captureException = jest.fn();
+    // Mock the `configuration` module to fake an `apiSentryDsn`.
+    jest.mock("../../configuration", () => ({
+      configuration: { apiSentryDsn: "some sentry dsn" }
+    }));
     // Only now require the logger module so it uses the mocks.
     logger = require("./").logger;
   });
@@ -30,6 +37,8 @@ describe("test the output of the logger", () => {
     expect(log).toContain(now);
     // Contains the message.
     expect(log).toContain("just an info");
+    // Sentry.captureException shouldn't have been called.
+    expect(sentry.captureException).not.toHaveBeenCalled();
   });
   test("error uses console.log and includes the stack trace", () => {
     const error = new Error("this is an error");
@@ -46,5 +55,7 @@ describe("test the output of the logger", () => {
     expect(log).toContain("this is an error");
     // ... and contains the stack trace.
     expect(log).toContain(JSON.stringify(error.stack));
+    // Sentry.captureException should have been called with the error.
+    expect(sentry.captureException).toHaveBeenCalledWith(error);
   });
 });