+++ packages/api/CHANGELOG.md
@@ -8,7 +8,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-* **logger:** change logging strategy: switch from [winston](https://github.com/winstonjs/winston) and [raven](https://www.npmjs.com/package/raven) to [pino](http://getpino.io/) and [pino-sentry](https://github.com/aandrewww/pino-sentry/blob/master/README.md). See [PR #116](https://github.com/SocialGouv/egapro/pull/116)
+* **logger:** change logging strategy: switch from [winston](https://github.com/winstonjs/winston) and [raven](https://www.npmjs.com/package/raven) to [pino](http://getpino.io/) and [sentry](https://www.npmjs.com/package/@sentry/node). See [PR #116](https://github.com/SocialGouv/egapro/pull/116)